Added temporary debug call (mcbtime), changed index in evrTimeGetFifo to be 64 bits, added code to skip any queued entries when evrTimeEvent is first called.
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/event/current@9738 ececa60f-9f63-0410-97c7-ea05888eb260

+++ evrSupport/src/evrTime.h
@@ -135,10 +135,10 @@ int evrTimeGetFromEdef    (unsigned int     edefIdx,
                            epicsEnum16     *edefSevr_p);
 int evrTimeGet            (epicsTimeStamp  *epicsTime_ps,
                            unsigned int     eventCode);
-int evrTimeGetFifo        (epicsTimeStamp  *epicsTime_ps,
-                           unsigned int     eventCode,
-                           unsigned int    *idx,
-                           int              incr);
+int evrTimeGetFifo        (epicsTimeStamp     *epicsTime_ps,
+                           unsigned int        eventCode,
+                           unsigned long long *idx,
+                           int                 incr);
 int evrTimePutPulseID     (epicsTimeStamp  *epicsTime_ps,
                            unsigned int     pulseID);
 /* Routines used only by event module and Mpg application */
